backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
authlib
python-jose[cryptography]
pymongo
itsdangerous
httpx
openai
pdfplumber
python-docx
python-multipart
langchain
langgraph
langchain-openai
deepgram-sdk
mediapipe
opencv-python<4.10.0
deepface
tf-keras
numpy<1.28.0
